+++ core/terraform/destroy.sh
@@ -5,6 +5,14 @@ set -o pipefail
 set -o nounset
 # set -o xtrace
 
+# add trap to remove kv network exception
+# shellcheck disable=SC1091
+trap 'source "../../devops/scripts/kv_remove_network_exception.sh"' EXIT
+
+# now add kv network exception
+# shellcheck disable=SC1091
+source "../../devops/scripts/kv_add_network_exception.sh"
+
 # These variables are loaded in for us
 # shellcheck disable=SC2154
 ../../devops/scripts/terraform_wrapper.sh -g "${TF_VAR_mgmt_resource_group_name}" \